Remove pre-rebac code paths
