+++ app/web/src/components/layout/navbar/ChangeSetPanel.vue
@@ -43,11 +43,10 @@
           !selectedChangeSetName ||
           changeSetsStore.headSelected ||
           changeSetsStore.creatingChangeSet ||
-          (featureFlagsStore.REBAC &&
-            changeSetsStore.selectedChangeSet?.status ===
-              ChangeSetStatus.NeedsApproval)
+          changeSetsStore.selectedChangeSet?.status ===
+            ChangeSetStatus.NeedsApproval
         "
-        @click="openApprovalFlowModal"
+        @click="openAbandonConfirmationModal"
       />
     </div>
 
@@ -87,12 +86,6 @@
         </Stack>
       </form>
     </Modal>
-
-    <ApprovalFlowModal
-      ref="approvalFlowModalRef"
-      votingKind="abandon"
-      @completeVoting="changeSetsStore.ABANDON_CHANGE_SET"
-    />
     <AbandonChangeSetModal ref="abandonModalRef" />
   </div>
 </template>
@@ -111,7 +104,6 @@ import {
 import { useChangeSetsStore } from "@/store/change_sets.store";
 import { ChangeSetStatus } from "@/api/sdf/dal/change_set";
 import { useAuthStore } from "@/store/auth.store";
-import ApprovalFlowModal from "@/components/ApprovalFlowModal.vue";
 import AbandonChangeSetModal from "@/components/AbandonChangeSetModal.vue";
 import { useFeatureFlagsStore } from "@/store/feature_flags.store";
 
@@ -137,20 +129,12 @@ const changeSetDropdownOptions = computed(() => {
 const router = useRouter();
 const route = useRoute();
 
-const approvalFlowModalRef = ref<InstanceType<typeof ApprovalFlowModal> | null>(
-  null,
-);
-
 const abandonModalRef = ref<InstanceType<typeof AbandonChangeSetModal> | null>(
   null,
 );
 
-const openApprovalFlowModal = () => {
-  if (!featureFlagsStore.REBAC) {
-    approvalFlowModalRef.value?.open();
-  } else {
-    abandonModalRef.value?.open();
-  }
+const openAbandonConfirmationModal = () => {
+  abandonModalRef.value?.open();
 };
 
 const createModalRef = ref<InstanceType<typeof Modal>>();
@@ -187,17 +171,6 @@ function onSelectChangeSet(newVal: string) {
   }
 
   if (newVal && route.name) {
-    // do not allow people to navigate to a change set that NeedsApproval
-    // unless they were the one that initiated the merge request (avoids dead end)
-    if (
-      changeSetsStore.changeSetsById[newVal]?.status !== ChangeSetStatus.Open &&
-      changeSetsStore.changeSetsById[newVal]?.mergeRequestedByUserId !==
-        authStore.user?.pk &&
-      !featureFlagsStore.REBAC
-    ) {
-      return;
-    }
-
     // keep everything in the current route except the change set id
     // note - we use push here, so there is a new browser history entry
     let name = route.name;
