+++ bin/si-api-test/tests/6-get_head_changeset.ts
@@ -2,9 +2,14 @@ import assert from "node:assert";
 import { SdfApiClient } from "../sdf_api_client.ts";
 
 export default async function get_head_changeset(sdfApiClient: SdfApiClient) {
-  const data = await sdfApiClient.call({ route: "open_change_sets" });
+  const workspaceId = sdfApiClient.workspaceId;
+  const data = await sdfApiClient.call({
+    route: "open_change_sets", routeVars: {
+      workspaceId,
+    }
+  });
 
-  assert(data.headChangeSetId, "Expected headChangeSetId");
-  const head = data.changeSets.find((c) => c.id === data.headChangeSetId);
+  assert(data.defaultChangeSetId, "Expected headChangeSetId");
+  const head = data.changeSets.find((c) => c.id === data.defaultChangeSetId);
   assert(head, "Expected a HEAD changeset");
 }
