+++ lib/dal/src/change_set.rs
@@ -798,88 +798,6 @@ impl ChangeSet {
         Utc::now().format("%Y-%m-%d-%H:%M").to_string()
     }
 
-    pub async fn merge_vote(&mut self, ctx: &DalContext, vote: String) -> ChangeSetResult<()> {
-        let user_id = Self::extract_userid_from_context(ctx).await;
-        WsEvent::change_set_merge_vote(ctx, self.id, user_id, vote)
-            .await?
-            .publish_on_commit(ctx)
-            .await?;
-
-        Ok(())
-    }
-    pub async fn abandon_vote(&mut self, ctx: &DalContext, vote: String) -> ChangeSetResult<()> {
-        let user_id = Self::extract_userid_from_context(ctx).await;
-        WsEvent::change_set_abandon_vote(ctx, self.id, user_id, vote)
-            .await?
-            .publish_on_commit(ctx)
-            .await?;
-
-        Ok(())
-    }
-
-    pub async fn cancel_abandon_approval_flow(&mut self, ctx: &DalContext) -> ChangeSetResult<()> {
-        self.update_status(ctx, ChangeSetStatus::Open).await?;
-        let user_id = Self::extract_userid_from_context(ctx).await;
-        WsEvent::change_set_cancel_abandon_approval_process(ctx, self.id, user_id)
-            .await?
-            .publish_on_commit(ctx)
-            .await?;
-
-        Ok(())
-    }
-    pub async fn begin_abandon_approval_flow(&mut self, ctx: &DalContext) -> ChangeSetResult<()> {
-        self.update_status(ctx, ChangeSetStatus::NeedsAbandonApproval)
-            .await?;
-        let user_id = Self::extract_userid_from_context(ctx).await;
-        WsEvent::change_set_begin_abandon_approval_process(ctx, self.id, user_id)
-            .await?
-            .publish_on_commit(ctx)
-            .await?;
-        WsEvent::change_set_abandon_vote(
-            ctx,
-            ctx.visibility().change_set_id,
-            user_id,
-            "Approve".to_string(),
-        )
-        .await?
-        .publish_on_commit(ctx)
-        .await?;
-        Ok(())
-    }
-
-    pub async fn begin_approval_flow(&mut self, ctx: &DalContext) -> ChangeSetResult<()> {
-        self.update_status(ctx, ChangeSetStatus::NeedsApproval)
-            .await?;
-        let user_id = Self::extract_userid_from_context(ctx).await;
-        if let Some(user_pk) = user_id {
-            self.update_merge_requested_by_user_id(ctx, user_pk).await?;
-        }
-        WsEvent::change_set_begin_approval_process(ctx, self.id, user_id)
-            .await?
-            .publish_on_commit(ctx)
-            .await?;
-        WsEvent::change_set_merge_vote(
-            ctx,
-            ctx.visibility().change_set_id,
-            user_id,
-            "Approve".to_string(),
-        )
-        .await?
-        .publish_on_commit(ctx)
-        .await?;
-        Ok(())
-    }
-
-    pub async fn cancel_approval_flow(&mut self, ctx: &DalContext) -> ChangeSetResult<()> {
-        self.update_status(ctx, ChangeSetStatus::Open).await?;
-        let user_id = Self::extract_userid_from_context(ctx).await;
-        WsEvent::change_set_cancel_approval_process(ctx, self.id, user_id)
-            .await?
-            .publish_on_commit(ctx)
-            .await?;
-        Ok(())
-    }
-
     pub async fn abandon(&mut self, ctx: &DalContext) -> ChangeSetResult<()> {
         self.update_status(ctx, ChangeSetStatus::Abandoned).await?;
         let user_id = Self::extract_userid_from_context(ctx).await;
