+++ lib/sdf-server/src/service/change_set.rs
@@ -1,7 +1,7 @@
 use axum::{
     http::StatusCode,
     response::{IntoResponse, Response},
-    routing::{get, post},
+    routing::post,
     Router,
 };
 use dal::{
@@ -20,14 +20,8 @@ use crate::AppState;
 use super::ApiError;
 
 pub mod abandon_change_set;
-mod abandon_vote;
 pub mod add_action;
-pub mod apply_change_set;
-mod begin_abandon_approval_process;
-mod begin_approval_process;
 pub mod create_change_set;
-pub mod list_open_change_sets;
-mod merge_vote;
 mod rebase_on_base;
 mod status_with_base;
 
@@ -96,41 +90,15 @@ impl IntoResponse for ChangeSetError {
 
 pub fn routes() -> Router<AppState> {
     Router::new()
-        .route(
-            "/list_open_change_sets",
-            get(list_open_change_sets::list_open_change_sets),
-        )
         .route("/add_action", post(add_action::add_action))
         .route(
             "/create_change_set",
             post(create_change_set::create_change_set),
         )
         .route(
-            "/apply_change_set",
-            post(apply_change_set::apply_change_set),
-        )
-        .route(
             "/abandon_change_set",
             post(abandon_change_set::abandon_change_set),
         )
-        .route(
-            "/begin_approval_process",
-            post(begin_approval_process::begin_approval_process),
-        )
-        .route(
-            "/cancel_approval_process",
-            post(begin_approval_process::cancel_approval_process),
-        )
-        .route("/merge_vote", post(merge_vote::merge_vote))
-        .route(
-            "/begin_abandon_approval_process",
-            post(begin_abandon_approval_process::begin_abandon_approval_process),
-        )
-        .route(
-            "/cancel_abandon_approval_process",
-            post(begin_abandon_approval_process::cancel_abandon_approval_process),
-        )
-        .route("/abandon_vote", post(abandon_vote::abandon_vote))
         .route("/rebase_on_base", post(rebase_on_base::rebase_on_base))
         .route(
             "/status_with_base",
