--- lib/sdf-server/src/service/change_set/list_open_change_sets.rs
@@ -1,18 +1,0 @@
-use axum::Json;
-use dal::change_set::view::OpenChangeSetsView;
-
-use super::ChangeSetResult;
-use crate::extract::{AccessBuilder, HandlerContext};
-
-pub type ListOpenChangeSetsResponse = OpenChangeSetsView;
-
-pub async fn list_open_change_sets(
-    HandlerContext(builder): HandlerContext,
-    AccessBuilder(access_builder): AccessBuilder,
-) -> ChangeSetResult<Json<ListOpenChangeSetsResponse>> {
-    let ctx = builder.build_head(access_builder).await?;
-
-    let view = OpenChangeSetsView::assemble(&ctx).await?;
-
-    Ok(Json(view))
-}
